OTTER-551: address review feedback on three-option selector
Code-review-client changes:
- Rename isDecisionTerminal → hasDecision; the gate no longer implies a
  terminal workflow state now that needs-clarification routes the code
  back instead of ending it.
- Replace the nested Text component="span" inside the reject option's
  subtext with a plain <strong>; semantically honest and avoids the
  span-inside-span rendering.
- Humanize the labName fallback ('the research lab' instead of falling
  back to the raw org slug, which read as machine-y in user copy).
- Drop the handleDecisionChange wrapper around decision.onSelect; the
  EditableBody prop now passes the hook setter directly.
- Add a test asserting CodeReviewClient renders null when the
  collaboration feature flag is off, covering the early-return path
  that the rest of the suite assumed enabled.

Decision type extraction:
- Move the shared Decision union and toReviewDecision helper out of
  src/lib/proposal-review.ts into a new src/lib/review-decision.ts.
  proposal-review.ts now only holds proposal-specific state helpers
  (isSubmittedProposalReviewStatus, FEEDBACK_* word limits). The
  Decision type is used in code-review too, and importing it from
  '@/lib/proposal-review' inside code-review-client.tsx read as a lie
  about the boundaries.
