+++ src/hooks/use-proposal-review-mutation.ts
@@ -6,7 +6,7 @@ import { useRouter } from 'next/navigation'
 import { useMutation, useQueryClient } from '@/common'
 import { reportMutationError } from '@/components/errors'
 import { useProposalCollaborationFeatureFlag } from '@/components/openstax-feature-flag'
-import type { Decision } from '@/lib/proposal-review'
+import type { Decision } from '@/lib/review-decision'
 import { Routes } from '@/lib/routes'
 import { type SubmissionEvent } from '@/hooks/use-submission-redirect-listener'
 import { useReviewFeedbackProvider } from '@/lib/realtime/review-feedback-provider-context'
