+++ src/hooks/use-review-decision.ts
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import type { Decision } from '@/lib/proposal-review'
+import type { Decision } from '@/lib/review-decision'
 
 export function useReviewDecision() {
     const [selected, setSelected] = useState<Decision | null>(null)
